chore: Simplify
